pyproject.toml
[tool.poetry]
name = "django-rest-api"
version = "0.1.0"
description = "A Django REST API project with authentication and Swagger documentation"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
django-admin-interface = "^0.26.0"
django-colorfield = "^0.9.0"
python-dotenv = "^1.0.1"
django-jazzmin = "^2.6.0"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 